fead: improve repositories

+++ src/Net.Shared.Persistence.Abstractions/Entities/IPersistentProcess.cs
@@ -5,7 +5,7 @@ public interface IPersistentProcess : IPersistent
     Guid? HostId { get; set; }
     int StatusId { get; set; }
     int StepId { get; set; }
-    byte Attempt { get; set; }
+    int Attempt { get; set; }
     string? Error { get; set; }
     DateTime Updated { get; set; }
 }
